List the valid keywords from `finder-known-keywords'
The function for pointing out the invalid keywords used to send the user
to look up the valid keywords from `finder-known-keywords'. But why have
this extra step, why not just tell the user what the valid keywords are?

This is my attempt at doing just that, pointing out the invalid keyword
and then listing the valid keywords that the user can chose from in
order to pass the linting.