Fix bug when a package appears in multiple archives
In the previous version of the code, best-version contains the lowest
version and because package-lint raises an error if this lowest
version is lower than the requested version, there will always be an
error if a package appears with a low version in any archive. This
happens for json-mode: it exists in version 0.1 in gnu and 1.7.0 in
melpa-stable. Any request for a version greater than 0.1 will fail.