Silence Emacs 25's byte-compiler
`flymake-diag-region' and `flymake-make-diagnostic' were added in
Emacs 26.  `package-lint-setup-flymake' errors out if the user tries
to use the functionality provided by this library when using an older
Emacs, but that doesn't prevent warnings when compiling this library,
which happens even on incompatible Emacsen.

Declare the functions to silence the byte-compiler.  An alternative
approach would be to distribute this library as a separate package,
but that seems a bit drastic.