Speed up the linting process
Profiling reveals that most of the CPU time is spent in syntax-ppss, called by
package-lint--inside-comment-or-string-p, called from
package-lint--check-version-regexp-list. To speed it up, don't check if we're
inside a comment or a string until absolutely necessary, i.e. right before
reporting an error.

The times for a single package-lint-buffer call on a 300 KB emacs lisp source
file (namely, the latest helm.el) on my Haswell box are:

Before #137: 0.2 s
After #137: 2.05 s
This commit: 0.35 s

So this change clearly brings performance back to almost pre-#137 levels
without changing the semantics.