Add support for allowed package-to-symbol prefix mappings (#177)
Valid mappings are stored in
`package-lint--allowed-prefix-mappings`. Its default value is:

    '(("ox-" . ("org-"))
      ("ob-" . ("org-")))

This means that in packages starting with "ox-" and "ob-", functions
and symbols starting with "org-" are allowed. After the mapping, the
symbols still must match the whole package match. For example, in
package "ox-foobar", symbols starting with "org-foobar" will be
allowed.

Fixes #89.