Replace certain single quotes in docstrings
The byte-compiler recently got more fussy about quotes.

This was rebased on top of another attempt to address this, which I
find weird.  IMO the two options are \\='symbol and `symbol', and I
prefer the latter.  I have been using that all along, even while
'symbol did not result in a warning yet.