Remove face eldoc-highlight-function-argument from 25.1/removed
This is a face.
Doc: "Face used for the argument at point in a function's argument
list.  Note that this face has no effect unless the
`eldoc-documentation-function' handles it explicitly."

I checked Emacs 25, 25.1, 26 and Master
25:
https://github.com/emacs-mirror/emacs/blob/emacs-25/lisp/emacs-lisp/eldoc.el#L113
25.1:
https://github.com/emacs-mirror/emacs/blob/f0eb70d8935be90f7c03e187c12d9b60e7214cc6/lisp/emacs-lisp/eldoc.el#L113
26:
https://github.com/emacs-mirror/emacs/blob/emacs-26/lisp/emacs-lisp/eldoc.el#L113
Master:
https://github.com/emacs-mirror/emacs/blob/master/lisp/emacs-lisp/eldoc.el#L141

The face is present in all of them.  This face is integral to the
eldoc ux, so a removal and re-introduction seems unlikely.