Ability to specify a prefix for package definitions
By default, this is implicitly derived from the buffer being linted,
but as this would signal warnings for multi-file packages, there is
also already the ability to specify the "main" file for a package,
which is then used to derive the common prefix for all package
modules. Yet, this too is inadequate for component packages in a
package suite that share a common prefix.

Explicitly specifying the prefix affords the linter a general way to
check such cases.