Add tests for `package-lint-main-file`
Testing this functionality requires setting up a main file buffer
separate from the buffer being linted. Formerly, these two facilities
were coupled in the core `package-lint-test--run` utility. This
refactors that utility so that it is a thin wrapper around a dedicated
helper for setting up a test buffer with boilerplate such as a header,
commentary, etc.

This refactored helper is used in the new test for deriving the lint
prefix from the declared "main file," functionality which formerly was
untested.